refactor: move tracker to src root
